[REF] web: convert view service to plugin
In this commit, we rewrite the view service as a plugin. For legacy
purposes, we keep the view service (as a service). We also adapt the
codebase to reflect the changes on the service.
