+++ odoo/upgrade_code/owl3-migration.py
@@ -1128,6 +1128,7 @@ SERVICES_MAPPING = {
     "sortable": ['SortablePlugin', '@web/core/util/sortable_plugin'],
     "simple_notification": ['SimpleNotificationPlugin', '@bus/simple_notification_plugin'],
     "ui": ['UIPlugin', '@web/core/ui/ui_plugin'],
+    "view": ['ViewPlugin', '@web/views/view_plugin'],
     "worker_service": ['WorkerPlugin', '@bus/services/worker_plugin'],
 }
